Newest version of the generic script.
This is a FreeCAD script to create spacers to reinforce concrete piles. The default values are chosen for a pile made from 110-mm sewage pipe, with 4 rebars along the pile. The diameters of the rebar holders shall be chosen according to the available reinforcement bars, and shall be slightly smaller than the external diameter of the bar.